+++ src/components/terminal/web_shell_prompt.svelte
